// src/app/core/interfaces/tableHead.model.ts
export interface TableHeadInterface {
    name: string;
    label: string;
    type?: string;
    format?: string;
    prefix?: string;
    decimals?: number;
    event?: string;
    orientacion?: string;
    wrap?: number;
    title?: string;
    short?: boolean;
    shorten?: number;
    colsNames?: any;
    colsText?: any;
    colsTextTitle?: any;
    // con_suspension?: boolean;
    chek?: string;
    colsColor?: string;
}